simplify UI.Layout
no need to pass the entire rectangle of where a child will be layed
out. just the size matters. makes it easier to implement.

same for the "cur" point. just ask the child to lay itself out at
the start of a new line. if it turns out to fit on the previous
line, place it there. child doesn't have to know about it.