welcome vi editing in edit windows
i've wanted this for a very long time.
now i'm getting it as a side effect of a ui lib. (:

still lots to do.
this is quick 'n dirty code. should be cleaned up, i've glanced at vixen, which has shorter code for key handling, we should be able to get that too.
btw, we don't strive to be a full vi clone. shouldn't be too hard though. i'll add features as i'm missing them.

major missing feature is undo/redo now, and vi search, and vi repeat last command, and safely overwriting a file, and not doing i/o in main loop.