lots of rework for better performance
we now track needs-draw and needs-layout state in each Kid.
and we pass that Kid around a lot.
this allows us to know which UI's need layout/draw, so we can do far fewer draws: only there where it is needed.

there are plenty of other changes in here, and plenty of new bugs.