add DUIOpts to NewDUI params, and take away the dimensions parameter
dimensions is now in DUIOpts.
it isn't very useful to specify, we store the actual size on disk
for the app after resize anyway. only the first run needs to be
reasonable in size. 800x600 is fine for now. this means most apps
have to specify one less thing to create a new ui.