don't panic/quit on errors during ui operation
the typical cases that could go wrong are/were:
- allocating an image to render on
- reading from a file to display text in an Edit

the main loop is now slightly different. there is no longer a
DUI.Done channel. that has become a DUI.Error channel. it is still
closed when the window goes away.  but in case of error, an error
is sent there. the app can determine what to do: log it to stderr,
show it in the UI, quit.

Edit.Error can be set to a channel. if set, it will receive the
error instead of DUI.Error.  if not set, the error goes to DUI.Error.